chore: merge spatial filter modules
